requirements.txt
Flask
Flask-SQLAlchemy
psycopg2
pusher